Fix bug where message_type header was required for NATS even with DefaultIncomingMessage
- Updated CoreNatsSubscriber and JetStreamSubscriber to allow messages without 'message-type' header if the endpoint has a defined MessageType (configured via DefaultIncomingMessage).
- Added debug logging to JetStreamSubscriber for skipped messages, consistent with CoreNatsSubscriber.
- Added a regression test in NatsTransportIntegrationTests.cs verifying that raw NATS messages without headers are correctly received when a default message type is set.
- Ensured consistency with other transports like RabbitMQ which do not have this restriction.
